AR-66 AR-68
Splitting pin and device endpoints
