+++ app/__init__.py
@@ -9,6 +9,8 @@ from flask_cors import CORS
 from app.routes.main import main_bp
 from app.routes.demo_crud import crud_bp
 from app.routes.rpi import rpi_bp
+from app.routes.rpi_pin import rpi_pin_bp
+from app.routes.rpi_device import rpi_device_bp
 from app.services.db import DBManager
 from config import Config
 
@@ -41,6 +43,8 @@ def create_app():
     api.register_blueprint(crud_bp)
     if os.getenv("RPI_MODULE"):
         api.register_blueprint(rpi_bp)
+        api.register_blueprint(rpi_pin_bp)
+        api.register_blueprint(rpi_device_bp)
         sqlite_manager = DBManager()
         sqlite_manager.reset_db_settings("sqlite-rpi")
         sqlite_manager.check_coherence()
